+++ scripts/generate-app-config/importSideEffects.js
@@ -3,16 +3,28 @@ const path = require('path');
 const ImportAggregator = require('./generation-objects/ImportAggregator');
 
 /**
+* Resolve relative paths that start with '.'
+*/
+const resolveRelativePaths = sideEffect => (sideEffect.charAt(0) === '.' ? path.resolve('dev-site-config', sideEffect) : sideEffect);
+
+/**
+* glob paths, warn if nothing is found for the glob.
+*/
+const globSideEffects = (acc, sideEffectGlob) => {
+  const result = glob.sync(`${sideEffectGlob}?(.*)`);
+  if (result.length === 0) {
+    // eslint-disable-next-line no-console
+    console.log(`[terra-dev-site] No Side Effect imports found at ${sideEffectGlob}`);
+  }
+  return acc.concat(result);
+};
+
+/**
 * Resolve relative paths that start with '.', then glob paths, then apply side effect imports
 */
-// Disabling linter to preserve more readable formatting
-/* eslint-disable */
 const importSideEffects = (sideEffects, imports) => {
-  sideEffects.map(sideEffectGlob =>
-    (sideEffectGlob.charAt(0) === '.' ? path.resolve('dev-site-config', sideEffectGlob) : sideEffectGlob)).reduce((acc, sideEffectGlob) =>
-      acc.concat(glob.sync(`${sideEffectGlob}?(.*)`)), []).forEach(sideEffectPath =>
-        imports.addImport(ImportAggregator.relativePath(sideEffectPath)));
+  const sideEffectFiles = sideEffects.map(resolveRelativePaths).reduce(globSideEffects, []);
+  sideEffectFiles.forEach(sideEffectPath => imports.addImport(ImportAggregator.relativePath(sideEffectPath)));
 };
-/* eslint-enable */
 
 module.exports = importSideEffects;