+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Log warning if side-effects globs are not resolved to something.
 
 4.0.0 - (April 12, 2019)
 ----------