+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * headHtml added to config.
 
+### Changed
+* Log warning if side-effects globs are not resolved to something.
+
 4.0.0 - (April 12, 2019)
 ----------
 ### Changed